User may now specify toolset through CMake GUI
The -T parameter to CMake may now be specified through QtDialog (cmake-gui)
via a new text field in the first-time configure wizard (below the generator
chooser).

The generator factories specify whether or not they support toolsets. This
information is propagated to the QT code and used to determine if the
selected generator should also display the optional Toolset widgets.